feat(platform): MACOS-PORT-1 - native Apple Silicon build to the GL 4.3 wall
First macOS/arm64 port of the engine. It builds, links, and runs natively (SDL2
cocoa driver, window, GPU-cull selftest passes) up to SDL_GL_CreateContext for
an OpenGL 4.3 core context, which macOS native GL (4.1 cap) rejects. Zink plus
MoltenVK is the next step; base game only (assimp FBX mech import off).

The Linux port already did the D3D-to-GL and Win32-to-SDL2 work, so the macOS
delta is mechanical (39 files, +411/-28):

- x86 asm to ARM64: rdtsc uses cntvct_el0 (gameos.hpp), stack walk uses
  __builtin_return_address (heap.cpp)
- LP64 macOS has unsigned long distinct from unsigned long long: readIdULong
  overload (inifile), MemoryStreamIO overloads (memorystream.hpp), and elfHash
  forced to a 32-bit accumulator so FST asset keys still match
- MSVC-to-POSIX name shims in platform_str.h and platform_stdlib.h (_stricmp,
  _strnicmp, _snprintf, _vsnprintf, _alloca, _putenv_s, sprintf_s,
  OutputDebugString)
- Win32 file/time APIs implemented on POSIX in platform_winbase.{h,cpp}
  (QueryPerformanceCounter/Frequency, GetFileAttributesEx, GetFileAttributesA,
  GetFullPathName, CompareFileTime, CreateFile/ReadFile/CloseHandle,
  GetTickCount64, GetCurrentThreadId), so mod-discovery works natively
- compat shims in 3rdparty/include (windows.h, intrin.h, process.h, malloc.h,
  and a fix to unistd.h) that include_next on Windows and forward to the
  emulation layer or system headers elsewhere
- assimp off: stub mc2mechanim::Imported* in assimp_importer.cpp so the engine
  links without the importer, and always-compile that TU
- FindFFmpegLGPL.cmake: Apple branch links Homebrew ffmpeg dylibs
- Windows-only bits guarded: launcher relaunch (gameosmain.cpp), SEH
  (ablexec.cpp), wide-char _wfopen and MoveFileExW, DLL-copy POST_BUILD steps,
  Windows tool subdirs

missionName gets an owning definition in logisticsmissioninfo.cpp because
code/logmain.cpp is not in the build. Every change is tagged // macos-port:.

+++ code/logisticsdialog.cpp
@@ -633,7 +633,7 @@ void LogisticsSaveDialog::initDialog( const char* path, bool bCampaign )
 			LogisticsSaveDialog* self;
 			unsigned long itemColor;
 		};
-		ModCampaignCtx ctx = { this, edits[0].getColor() };
+		ModCampaignCtx ctx = { this, (unsigned long)edits[0].getColor() }; // macos-port: clang rejects long->unsigned long narrowing in {}
 		EnumerateModCampaignFiles([](const char* fitFileName, const char* fullPath, void* ud) {
 			ModCampaignCtx* c = static_cast<ModCampaignCtx*>(ud);
